requirements.txt
pandas
numpy
scikit-learn
plotly
matplotlib
seaborn
lightgbm
xgboost
dill
flask
flask-wtf
flask-bootstrap
flask-mail
flask-moment
flask_sqlalchemy
psycopg2
flask_fontawesome
Flask-Migrate
ensure
python-box
pyYAML
python-dotenv
jupyter


